Added CRUD functionality for User and Appointment through the MgmtService.
Changed Appointment entity types for start/end time from Time to LocalDateTime.

Added jackson jsr310 module to handle appointment entity change. Set the object mapper in MessageConverter in JMSConfig to support this. Added dependency for this module in POM files. Tested all CRUD functionality with postman (reach out for API calls).

Deleting has a foreign key constraint issue if it's present in the join table; likely need to set a cascade value to fix. If a resource is not found, the MgmtService never gets an HTTP response, needs to be fixed (timeout?).